Adopt docker to run custom container
