+++ src/template_generation/file_system_agent/requirements.txt
@@ -1,0 +1,1 @@
+Flask==3.0.2
